requirements.txt
absl-py==2.3.1
annotated-types==0.7.0
anthropic==0.75.0
anyio==4.12.0
backoff==2.2.1
bert-score==0.3.13
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
contourpy==1.3.2
cycler==0.12.1
distro==1.9.0
docstring_parser==0.17.0
dydantic==0.0.8
exceptiongroup==1.3.1
faiss-cpu==1.13.1
filelock==3.20.0
fonttools==4.61.0
fsspec==2025.12.0
greenlet==3.3.0
grpcio==1.76.0
h11==0.16.0
h2==4.3.0
hf-xet==1.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
hyperframe==6.1.0
idna==3.11
Jinja2==3.1.6
jiter==0.12.0
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.9
langchain==1.1.2
langchain-anthropic==1.2.0
langchain-core==1.1.1
langchain-openai==1.1.0
langgraph==1.0.4
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.5
langgraph-sdk==0.2.14
langmem==0.0.30
langsmith==0.4.56
MarkupSafe==3.0.3
matplotlib==3.10.7
mem0ai==1.0.1
modelscope==1.33.0
mpmath==1.3.0
networkx==3.4.2
nltk==3.9.2
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
openai==2.9.0
orjson==3.11.5
ormsgpack==1.12.0
packaging==25.0
pandas==2.3.3
pillow==12.0.0
portalocker==3.2.0
posthog==7.0.1
protobuf==5.29.5
pydantic==2.12.5
pydantic_core==2.41.5
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
qdrant-client==1.16.1
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
rouge_score==0.1.2
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.15.3
sentence-transformers==5.1.2
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.1
torch==2.9.1
tqdm==4.67.1
transformers==4.57.3
triton==3.5.1
trustcall==0.0.39
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.6.0
uuid_utils==0.12.0
xxhash==3.6.0
zep-cloud==3.13.0
zstandard==0.25.0
